backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
xgboost
joblib
huggingface_hub
python-dotenv
numpy